# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.3
info:
  title: B-ting Backend API
  description: B-ting backend API documentation.
  version: 0.0.1-SNAPSHOT
servers:
  - url: https://api.buting.store
    description: Production server
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/travels/{travelId}/records/{travelRecordId}:
    get:
      tags:
        - TravelRecord
      summary: 여행 기록 초안 상세 조회
      description: 초안 상세를 조회합니다. 원본 여행에 속한 작성자만 조회할 수 있습니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelId"
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    patch:
      tags:
        - TravelRecord
      summary: 여행 기록 초안 수정
      description: 초안의 제목, 본문, 커버 이미지를 수정합니다. 요청 본문은 생략할 수 있습니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelId"
        - $ref: "#/components/parameters/TravelRecordId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelRecordUpdateRequest"
            example:
              title: 수정한 부산 여행
              content: 수정한 여행 기록 본문입니다.
              coverImageUrl: https://example.com/new-cover.jpg
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordResponse"
        "400":
          description: 검증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /api/v1/travel-records:
    get:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 피드 조회
      description: 공개 발행된 여행 기록 피드를 조회합니다. 커서 페이지네이션, 키워드/장소/여행일자/지역/도시 필터, 정렬을 지원합니다.
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: 이전 페이지 응답의 nextCursor 값입니다.
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 제목, 본문, 장소명을 검색합니다.
        - name: placeId
          in: query
          required: false
          description: 특정 장소를 포함한 여행 기록만 조회할 때 사용하는 장소 ID입니다.
          schema:
            type: string
          example: google-place-id
        - name: travelStartDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: travelEndDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: region
          in: query
          required: false
          schema:
            type: string
          description: 장소명 또는 주소에 해당 지역 문자열이 포함된 여행 기록만 조회합니다.
          example: 부산
        - name: city
          in: query
          required: false
          schema:
            type: string
          description: 장소명 또는 주소에 해당 도시 문자열이 포함된 여행 기록만 조회합니다.
          example: 서귀포
        - name: sort
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TravelRecordFeedSort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordFeedPageResponse"
        "400":
          description: 잘못된 필터 또는 커서
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/{travelRecordId}:
    get:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 상세 조회
      description: 공개 발행된 여행 기록 상세를 조회합니다. 조회 시 조회수가 증가합니다.
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordResponse"
        "404":
          description: 공개 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/me:
    get:
      tags:
        - TravelRecord
      summary: 내 여행 기록 목록 조회
      description: 인증된 사용자가 작성한 DRAFT, PUBLISHED, HIDDEN 상태의 여행 기록 목록을 조회합니다.
      security:
        - opaqueToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TravelRecordManageResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/me/bookmarks:
    get:
      tags:
        - TravelRecord
      summary: 내가 북마크한 여행 기록 조회
      security:
        - opaqueToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TravelRecordBookmarkResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/me/{travelRecordId}:
    get:
      tags:
        - TravelRecord
      summary: 내 여행 기록 상세 조회
      description: 작성자는 DRAFT, PUBLISHED, HIDDEN 상태의 기록을 모두 조회할 수 있습니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    patch:
      tags:
        - TravelRecord
      summary: 내 여행 기록 수정
      description: DRAFT는 자유롭게 수정할 수 있고, PUBLISHED/HIDDEN은 제목, 본문, 커버 이미지를 수정할 수 있습니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelRecordUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordResponse"
        "400":
          description: 검증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/me/{travelRecordId}/hide:
    post:
      tags:
        - TravelRecord
      summary: 내 여행 기록 숨김 처리
      description: 작성자 본인의 공개 기록을 HIDDEN 상태로 변경하여 공개 피드와 공개 상세 조회에서 제외합니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "200":
          description: 숨김 처리 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/me/{travelRecordId}/republish:
    post:
      tags:
        - TravelRecord
      summary: 내 숨김 여행 기록 재공개
      description: 작성자 본인의 HIDDEN 상태 기록을 다시 PUBLISHED 상태로 변경합니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "200":
          description: 재공개 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordResponse"
        "400":
          description: 재공개할 수 없는 기록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/{travelRecordId}/bookmarks:
    post:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 북마크
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "201":
          description: 북마크 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordBookmarkResponse"
        "400":
          description: 중복 북마크 또는 공개되지 않은 기록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - TravelRecord
      summary: 여행 기록 북마크 삭제
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "204":
          description: 북마크 삭제 성공
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/{travelRecordId}/likes:
    post:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 좋아요
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "201":
          description: 좋아요 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordLikeResponse"
        "400":
          description: 중복 좋아요 또는 공개되지 않은 기록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - TravelRecord
      summary: 여행 기록 좋아요 취소
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "204":
          description: 좋아요 취소 성공
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/{travelRecordId}/comments:
    post:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 댓글 작성
      description: 공개 발행된 여행 기록에 댓글을 작성합니다. 댓글 내용은 앞뒤 공백을 제거해 저장합니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelRecordCommentCreateRequest"
            example:
              content: 이 코스 참고하기 좋아요!
      responses:
        "201":
          description: 댓글 작성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordCommentResponse"
        "400":
          description: 댓글 내용 검증 실패 또는 공개되지 않은 기록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 공개 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    get:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 댓글 목록 조회
      description: 공개 발행된 여행 기록의 댓글을 작성일 오름차순으로 조회합니다.
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
      responses:
        "200":
          description: 댓글 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TravelRecordCommentResponse"
        "404":
          description: 공개 여행 기록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/v1/travel-records/{travelRecordId}/comments/{commentId}:
    patch:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 댓글 수정
      description: 댓글 작성자만 본인 댓글을 수정할 수 있습니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
        - name: commentId
          in: path
          required: true
          description: 댓글 ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelRecordCommentUpdateRequest"
            example:
              content: 직접 따라가 봤는데 동선이 좋아요.
      responses:
        "200":
          description: 댓글 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRecordCommentResponse"
        "400":
          description: 댓글 내용 검증 실패 또는 공개되지 않은 기록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 댓글 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 공개 여행 기록 또는 댓글을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    delete:
      tags:
        - TravelRecord
      summary: 공개 여행 기록 댓글 삭제
      description: 댓글 작성자만 본인 댓글을 삭제할 수 있습니다.
      security:
        - opaqueToken: []
      parameters:
        - $ref: "#/components/parameters/TravelRecordId"
        - name: commentId
          in: path
          required: true
          description: 댓글 ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 댓글 삭제 성공
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: 댓글 작성자가 아님
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: 공개 여행 기록 또는 댓글을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  parameters:
    TravelId:
      name: travelId
      in: path
      required: true
      description: 여행 ID
      schema:
        type: string
        format: uuid
      example: 10000000-0000-0000-0000-000000000001
...
    TravelRecordId:
      name: travelRecordId
      in: path
      required: true
      description: 여행 기록 ID
      schema:
        type: string
        format: uuid
      example: 60000000-0000-0000-0000-000000000001
...
  securitySchemes:
    opaqueToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    TravelRecordUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          nullable: true
          example: 수정된 부산 여행
        content:
          type: string
          nullable: true
          description: 여행기 본문입니다. 필수는 아닙니다.
          example: 수정된 여행 기록 본문입니다.
        coverImageUrl:
          type: string
          maxLength: 1000
          nullable: true
          example: https://example.com/new-cover.jpg
        overallRating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          description: 여행 플랜 전체 평점입니다.
          example: 4
    TravelRecordStatus:
      type: string
      enum:
        - DRAFT
        - PUBLISHED
        - HIDDEN
      example: PUBLISHED
    TravelRecordFeedSort:
      type: string
      enum:
        - LATEST
        - MOST_LIKED
        - MOST_VIEWED
      example: LATEST
    TravelRecordResponse:
      type: object
      properties:
        travelRecordId:
          type: string
          format: uuid
          example: 60000000-0000-0000-0000-000000000001
        originalTravelId:
          type: string
          format: uuid
          nullable: true
          example: 10000000-0000-0000-0000-000000000001
        authorId:
          type: string
          format: uuid
          example: 50000000-0000-0000-0000-000000000001
        title:
          type: string
          nullable: true
          example: 부산 맛집 여행
        content:
          type: string
          nullable: true
          example: 여유롭게 다녀온 2일 부산 여행 코스입니다.
        coverImageUrl:
          type: string
          nullable: true
          example: https://example.com/cover.jpg
        overallRating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          description: 여행 플랜 전체 평점입니다.
          example: 5
        travelStartDate:
          type: string
          format: date
          nullable: true
          example: "2026-08-01"
        travelEndDate:
          type: string
          format: date
          nullable: true
          example: "2026-08-02"
        status:
          $ref: "#/components/schemas/TravelRecordStatus"
        publishedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-08-03T10:00:00"
        likeCount:
          type: integer
          format: int64
          example: 7
        viewCount:
          type: integer
          format: int64
          example: 120
        days:
          type: array
          items:
            $ref: "#/components/schemas/TravelRecordDayResponse"
...
    TravelRecordFeedPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TravelRecordFeedResponse"
        nextCursor:
          type: string
          nullable: true
          example: TEFURVNUfDB8MjAyNi0wOC0wM1QxMDowMDowMHwyMDI2LTA4LTAzVDA5OjAwOjAw
        hasNext:
          type: boolean
          example: true
...
    TravelRecordManageResponse:
      type: object
      properties:
        travelRecordId:
          type: string
          format: uuid
          example: 60000000-0000-0000-0000-000000000001
        originalTravelId:
          type: string
          format: uuid
          nullable: true
          example: 10000000-0000-0000-0000-000000000001
        authorId:
          type: string
          format: uuid
          example: 50000000-0000-0000-0000-000000000001
        title:
          type: string
          nullable: true
          example: 부산 맛집 여행
        content:
          type: string
          nullable: true
          example: 여유롭게 다녀온 2일 부산 여행 코스입니다.
        coverImageUrl:
          type: string
          nullable: true
          example: https://example.com/cover.jpg
        overallRating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          description: 여행 플랜 전체 평점입니다.
          example: 5
        travelStartDate:
          type: string
          format: date
          nullable: true
          example: "2026-08-01"
        travelEndDate:
          type: string
          format: date
          nullable: true
          example: "2026-08-02"
        status:
          $ref: "#/components/schemas/TravelRecordStatus"
        publishedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-08-03T10:00:00"
        likeCount:
          type: integer
          format: int64
          example: 7
        viewCount:
          type: integer
          format: int64
          example: 120
        createdAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-08-03T09:00:00"
        updatedAt:
          type: string
          format: date-time
          nullable: true
          example: "2026-08-03T09:30:00"
    TravelRecordBookmarkResponse:
      type: object
      properties:
        bookmarkId:
          type: string
          format: uuid
          example: 80000000-0000-0000-0000-000000000001
        bookmarkedAt:
          type: string
          format: date-time
          example: "2026-08-03T11:00:00"
        travelRecord:
          $ref: "#/components/schemas/TravelRecordFeedResponse"
    TravelRecordLikeResponse:
      type: object
      properties:
        likeId:
          type: string
          format: uuid
          example: 81000000-0000-0000-0000-000000000001
        travelRecordId:
          type: string
          format: uuid
          example: 60000000-0000-0000-0000-000000000001
        likedAt:
          type: string
          format: date-time
          example: "2026-08-03T11:10:00"
        likeCount:
          type: integer
          format: int64
          example: 8
    TravelRecordCommentCreateRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          maxLength: 1000
          description: 댓글 내용입니다. 앞뒤 공백은 저장 전에 제거됩니다.
          example: 이 코스 참고하기 좋아요!
    TravelRecordCommentUpdateRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          maxLength: 1000
          description: 수정할 댓글 내용입니다. 앞뒤 공백은 저장 전에 제거됩니다.
          example: 직접 따라가 봤는데 동선이 좋아요.
    TravelRecordCommentResponse:
      type: object
      properties:
        commentId:
          type: string
          format: uuid
          example: 82000000-0000-0000-0000-000000000001
        travelRecordId:
          type: string
          format: uuid
          example: 60000000-0000-0000-0000-000000000001
        authorId:
          type: string
          format: uuid
          example: 50000000-0000-0000-0000-000000000001
        authorNickname:
          type: string
          example: 여행자
        authorProfileImageUrl:
          type: string
          nullable: true
          example: https://example.com/profile.jpg
        content:
          type: string
          example: 이 코스 참고하기 좋아요!
        createdAt:
          type: string
          format: date-time
          example: "2026-08-03T11:20:00"
        updatedAt:
          type: string
          format: date-time
          example: "2026-08-03T11:25:00"
...
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Email is required.
        data:
          nullable: true
...